requirements.txt
openai-whisper>=20231117
torch>=1.10.0
torchaudio>=0.10.0
numpy<2.3
ffmpeg-python>=0.2.0